// src/types/itemFetchArgument.type.ts
export interface WeaponArgument {
  distanceSelected: string | null;
  legalSelected: string | null;
  clSelected: string | null;
  gradeSelected: string | null;
  weaponSelected: string | null;
}

export interface MutantArmorArgument {
  genderSelected: string | null;
  clSelected: string | null;
  gradeSelected: string | null;
}